Move solar_angular_radius and coordinate_is_on_solar_disk to coordinates.utils
Moves coordinate-only functions from sunpy.map.maputils to sunpy.coordinates.utils
as requested in #8445. Maintains backwards compatibility by re-exporting from
sunpy.map.maputils.

Closes #8445
