+++ examples/computer_vision_techniques/mask_disk.py
@@ -9,7 +9,8 @@ import matplotlib.pyplot as plt
 
 import sunpy.map
 from sunpy.data.sample import AIA_171_IMAGE
-from sunpy.map.maputils import all_coordinates_from_map, coordinate_is_on_solar_disk
+from sunpy.coordinates.utils import coordinate_is_on_solar_disk
+from sunpy.map.maputils import all_coordinates_from_map
 
 ###############################################################################
 # We start with the sample data.
