+++ sunpy/net/jsoc/attrs.py
@@ -178,7 +178,7 @@ class Cutout(DataAttr):
         center = SkyCoord(center_x, center_y, frame=bottom_left.frame)
         if tracking:
             # import here so net won't depend on map
-            from sunpy.map.maputils import coordinate_is_on_solar_disk
+            from sunpy.coordinates.utils import coordinate_is_on_solar_disk
             if not coordinate_is_on_solar_disk(center):
                 raise ValueError("Tracking is enabled, but the center of the cutout "
                                  f"(Tx={center_x}, Ty={center_y}) is not on the solar disk.")
